Fix most of shadowed declarations and unused variables. Some remains because of EB or 2D/3D #ifdef.

+++ Source/NS_LES.cpp
@@ -96,7 +96,7 @@ NavierStokesBase::calc_mut_LES(MultiFab* mu_LES[BL_SPACEDIM], const Real time)
      tensorop.setLevelBC(0, &Uvel);
   }
               
-  const int dim_fluxes = pow(AMREX_SPACEDIM,2);
+  const int dim_fluxes = static_cast<int>(pow(AMREX_SPACEDIM,2));
   
   FluxBoxes fb(this,dim_fluxes);
   MultiFab** tensorflux = fb.get();
